+++ rapids-4-spark-tools/src/test/scala/com/nvidia/spark/rapids/tool/ToolTestUtils.scala
@@ -18,8 +18,13 @@ package com.nvidia.spark.rapids.tool
 
 import java.io.File
 
+import scala.collection.mutable.ArrayBuffer
+
+import com.nvidia.spark.rapids.tool.profiling.{ProfileArgs, ProfileUtils}
+
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.{DataFrame, SparkSession}
+import org.apache.spark.sql.rapids.tool.profiling.ApplicationInfo
 
 object ToolTestUtils extends Logging {
 
@@ -44,5 +49,20 @@ object ToolTestUtils extends Logging {
   def readExpectationCSV(sparkSession: SparkSession, path: String): DataFrame = {
     // make sure to change null value so empty strings don't show up as nulls
     sparkSession.read.option("header", "true").option("nullValue", "-").csv(path)
+  }
+
+  def processProfileApps(logs: Array[String],
+      sparkSession: SparkSession): ArrayBuffer[ApplicationInfo] = {
+    var apps: ArrayBuffer[ApplicationInfo] = ArrayBuffer[ApplicationInfo]()
+    val appArgs =
+      new ProfileArgs(logs)
+    var index: Int = 1
+    val eventlogPaths = appArgs.eventlog()
+    for (path <- eventlogPaths) {
+      apps += new ApplicationInfo(appArgs.numOutputRows.getOrElse(1000), sparkSession,
+        ProfileUtils.stringToPath(path)(0), index)
+      index += 1
+    }
+    apps
   }
 }
